Split ps_atoi_arr
I created a new function ps_arrlen() that finds the number of arguments
passed by the user. This number is used to allocate memory for the int
array. This functionality was included in the ps_atoi_arr() but since it
will be useful for allocating memory for the second stack I splitted the
function.