Add list manipulation code
I have changed the function that finds the maximum number, so that it
finds the minimum instead. The goal at this stage is to find the minimum
number in stack A, bring it to the top and push it to stack b. Once
completed stack B will contain all number sorted in a descended order.
Pushing them  back to stack A will result in a sorted stack in ascended
order.

To optimise the process, before bringing the min to the top, a function
will calculate the shortest path, and decide to use ra or rra.