Add linked list operations
I have updated the code to include several common operations on linked
lists. Added functions for rotation and reverse rotation of linked
lists. Finished the code that finds the fastest path to bring the
minimum number to the top of the stack and rotates the stack until the
minimum is on the top.

As a next step now, I must code the creation of stack B and pushing the
minimum value in stack B after getting them to the top of stack A.