Add quicksort and fn to find the median
I have added code for the quicksort algorithm and a function to find the
median of a sorted linked list. The idea is to sort the list find the
median and use the median to pre sort the stack A before using my own or
the insertion algorithm on it.